third_party/blink/web_tests/fast/events/touch/touch-input-element-change-documents-expected.txt

This test checks that we correctly update the touch event handler count when an Input element with default touch handlers changes documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
